+++ src/ruby/Mrubycfile
@@ -1,6 +1,6 @@
 ---
 test_dir: test
 test_tmp_dir: test/tmp
-mrubyc_src_dir: ../../lib/picoruby/build/repos/host/mruby-mrubyc/repos/mrubyc/src
+mrubyc_src_dir: ../../lib/picoruby/mrbgems/picoruby-mrubyc/repos/mrubyc/src
 mruby_lib_dir: app
 mrbc_path: ../../../../lib/picoruby/bin/picorbc # Relative path from test/tmp/
