+++ src/ruby/sig/keyboard.rbs
@@ -100,7 +100,7 @@ class Keyboard
   @ruby_mode: bool
   @via: VIA
   @debouncer: DebounceNone | DebouncePerRow | DebouncePerKey
-  @buffer: Buffer
+  @buffer: Terminal::Buffer
   @layer_changed_delay: Integer
   @partner_encoders: Array[RotaryEncoder]
   @macro_keycodes: Array[Integer]
